requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.9.0
Deprecated==1.2.14
deprecation==2.1.0
Flask==2.3.2
frozenlist==1.4.0
functions-framework==3.4.0
future==0.18.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.2.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
pydantic==2.2.0
pydantic_core==2.6.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
watchdog==3.0.0
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
